Python/AI_models/ATS_AI/requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
streamlit_lottie